+++ src/modules/index-header/index.tsx
@@ -1,51 +1,8 @@
-import React, { useEffect, useRef, useState } from "react";
-import Player from "@vimeo/player";
+import React from "react";
 import SubHeading from "../../components/ui/SubHeading";
 import Hero from "../../components/ui/Hero";
 import Button from "../../components/ui/Button";
-import ReactModal from "react-modal";
-import { LiteYoutubeEmbed } from "react-lite-yt-embed";
-
-const VimeoPlayer = ({ videoId }) => {
-  const playerRef = useRef(null);
-  const [isPlaying, setIsPlaying] = useState(false);
-  let player = null;
-
-  useEffect(() => {
-    player = new Player(playerRef.current, {
-      url: `https://vimeo.com/${videoId}`,
-    });
-
-    return () => {
-      player.off("play");
-    };
-  }, [videoId]);
-
-  const handlePlay = () => {
-    if (!isPlaying) {
-      player.play();
-      setIsPlaying(true);
-    }
-  };
-
-  return (
-    <div>
-      <div ref={playerRef} />
-      {!isPlaying && (
-        <div onClick={handlePlay}>
-          <div className="cursor-pointer play-container w-16 h-16 md:w-24 md:h-24 rounded-full bg-primary-500 flex justify-center items-center focus-visible:outline-none">
-            <img
-              src="/img/landing/play-icon.webp"
-              className="w-6 h-6 md:w-10 md:h-10"
-            />
-          </div>
-        </div>
-      )}
-    </div>
-  );
-};
-
-export default VimeoPlayer;
+import VimeoPlayer from "../../components/VimeoPlayer/VimeoPlayer";
 
 export const Header = () => {
   return (
